% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AlleleCodeBuf}
\alias{AlleleCodeBuf}
\title{Returns an empty two-row numeric matrix that ReadAlleles() can load to.}
\usage{
AlleleCodeBuf(pgen)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}
}
\value{
Numeric matrix with two rows, and appropriate number of columns for
ReadAlleles().
}
\description{
Returns an empty two-row numeric matrix that ReadAlleles() can load to.
}
