% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AppendBiallelic}
\alias{AppendBiallelic}
\title{Appends an unphased biallelic hardcall variant to a .pgen being written.}
\usage{
AppendBiallelic(pgen_writer, geno)
}
\arguments{
\item{pgen_writer}{Object returned by NewPgenWriter().}

\item{geno}{Integer or numeric vector of length sample_ct, with values in
\{0, 1, 2, NA\} indicating the number of ALT allele copies; -9 is also
accepted as a missing-value code.  (A buffer filled by ReadHardcalls() is
in this format.)}
}
\value{
No return value, called for side-effect.
}
\description{
Appends an unphased biallelic hardcall variant to a .pgen being written.
}
