% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BoolBuf}
\alias{BoolBuf}
\title{Returns a bool buffer that ReadAlleles() can load phasing information to.}
\usage{
BoolBuf(pgen)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}
}
\value{
Logical vector with appropriate length for ReadAlleles().
}
\description{
Returns a bool buffer that ReadAlleles() can load phasing information to.
}
