% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Buf}
\alias{Buf}
\title{Returns a numeric buffer that Read() or ReadHardcalls() can load to.}
\usage{
Buf(pgen)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}
}
\value{
Numeric vector with appropriate length for Read() and
ReadHardcalls().
}
\description{
Returns a numeric buffer that Read() or ReadHardcalls() can load to.
}
