% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ClosePgenWriter}
\alias{ClosePgenWriter}
\title{Finalizes a .pgen being written, and releases resources.  This must be
called after all variant_ct variants have been appended; the .pgen is not
valid before then.}
\usage{
ClosePgenWriter(pgen_writer)
}
\arguments{
\item{pgen_writer}{Object returned by NewPgenWriter().}
}
\value{
No return value, called for side-effect.
}
\description{
Finalizes a .pgen being written, and releases resources.  This must be
called after all variant_ct variants have been appended; the .pgen is not
valid before then.
}
