% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetAlleleCode}
\alias{GetAlleleCode}
\title{Look up an allele code.}
\usage{
GetAlleleCode(pvar, variant_num, allele_num)
}
\arguments{
\item{pvar}{Object returned by NewPvar().}

\item{variant_num}{Variant index (1-based).}

\item{allele_num}{Allele index (1-based).}
}
\value{
The allele_numth allele code for the variant_numth variant.
allele_num=1 corresponds to the REF allele, allele_num=2 corresponds to the
first ALT allele, allele_num=3 corresponds to the second ALT allele if it
exists and errors out otherwise, etc.
}
\description{
Look up an allele code.
}
