% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetAlleleCt}
\alias{GetAlleleCt}
\title{Returns max(2, <number of alleles the variant_numth variant is known to
have>).  Could be smaller than the actual number of alleles if no pvar was
provided to the NewPgen() call.}
\usage{
GetAlleleCt(pvar_or_pgen, variant_num)
}
\arguments{
\item{pvar_or_pgen}{Object returned by NewPvar() or NewPgen().}

\item{variant_num}{Variant index (1-based).}
}
\description{
Returns max(2, <number of alleles the variant_numth variant is known to
have>).  Could be smaller than the actual number of alleles if no pvar was
provided to the NewPgen() call.
}
