% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetMaxAlleleCt}
\alias{GetMaxAlleleCt}
\title{Returns max(2, <maximum number of alleles any variant is known to have>).
Could be smaller than the true number if no pvar was provided to the
NewPgen() call.}
\usage{
GetMaxAlleleCt(pvar_or_pgen)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}
}
\description{
Returns max(2, <maximum number of alleles any variant is known to have>).
Could be smaller than the true number if no pvar was provided to the
NewPgen() call.
}
