% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetRawSampleCt}
\alias{GetRawSampleCt}
\title{Returns the number of samples in the file.}
\usage{
GetRawSampleCt(pgen)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}
}
\value{
Number of samples.
}
\description{
Returns the number of samples in the file.
}
