% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetVariantChrom}
\alias{GetVariantChrom}
\title{Retrieve chromosome ID for given variant index.}
\usage{
GetVariantChrom(pvar, variant_num)
}
\arguments{
\item{pvar}{Object returned by NewPvar().}

\item{variant_num}{Variant index (1-based).}
}
\value{
Chromosome ID for the variant_numth variant.
}
\description{
Retrieve chromosome ID for given variant index.
}
