% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetVariantCt}
\alias{GetVariantCt}
\title{Returns the number of variants in the file.}
\usage{
GetVariantCt(pvar_or_pgen)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}
}
\description{
Returns the number of variants in the file.
}
