% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetVariantId}
\alias{GetVariantId}
\title{Returns the variant_numth variant ID.}
\usage{
GetVariantId(pvar, variant_num)
}
\arguments{
\item{pvar}{Object returned by NewPvar().}

\item{variant_num}{Variant index (1-based).}
}
\description{
Returns the variant_numth variant ID.
}
