% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetVariantPos}
\alias{GetVariantPos}
\title{Retrieve POS (base-pair coordinate on a chromosome) for given variant
index.}
\usage{
GetVariantPos(pvar, variant_num)
}
\arguments{
\item{pvar}{Object returned by NewPvar().}

\item{variant_num}{Variant index (1-based).}
}
\value{
POS for the variant_numth variant.
}
\description{
Retrieve POS (base-pair coordinate on a chromosome) for given variant
index.
}
