% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetWrittenVariantCt}
\alias{GetWrittenVariantCt}
\title{Returns the number of variants appended so far to a .pgen being written.}
\usage{
GetWrittenVariantCt(pgen_writer)
}
\arguments{
\item{pgen_writer}{Object returned by NewPgenWriter().}
}
\value{
Number of variants appended so far.
}
\description{
Returns the number of variants appended so far to a .pgen being written.
}
