% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HardcallPhasePresent}
\alias{HardcallPhasePresent}
\title{Returns true iff the file contains at least one phased heterozygous
hardcall.}
\usage{
HardcallPhasePresent(pgen)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}
}
\description{
Returns true iff the file contains at least one phased heterozygous
hardcall.
}
