% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HasSparseHardcalls}
\alias{HasSparseHardcalls}
\title{Returns whether hardcalls for the variant_numth variant and given allele
are represented in a sparse manner that is supported by
ReadSparseHardcalls().}
\usage{
HasSparseHardcalls(pgen, variant_num, allele_num = 2L)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}

\item{variant_num}{Variant index (1-based).}

\item{allele_num}{Allele index; 1 corresponds to REF, 2 to the first ALT
allele, 3 to the second ALT allele if it exists, etc.  Optional, defaults
to 2.}
}
\value{
True iff the (variant, allele) pair has a sparse representation
that can be returned by ReadSparseHardcalls().
}
\description{
Returns whether hardcalls for the variant_numth variant and given allele
are represented in a sparse manner that is supported by
ReadSparseHardcalls().
}
