% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IntBuf}
\alias{IntBuf}
\title{Returns an integer buffer that ReadHardcalls() can load to.}
\usage{
IntBuf(pgen)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}
}
\value{
Integer vector with appropriate length for ReadHardcalls().
}
\description{
Returns an integer buffer that ReadHardcalls() can load to.
}
