% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NewPgen}
\alias{NewPgen}
\title{Opens a .pgen or PLINK 1 .bed file.}
\usage{
NewPgen(filename, pvar = NULL, raw_sample_ct = NULL, sample_subset = NULL)
}
\arguments{
\item{filename}{.pgen/.bed file path.}

\item{pvar}{Object (see NewPvar()) corresponding to the .pgen's companion
.pvar; technically optional, but necessary for some functionality.}

\item{raw_sample_ct}{Number of samples in file; required if it's a PLINK 1
.bed file, otherwise optional.}

\item{sample_subset}{List of 1-based positions of samples to load;
optional, all samples are loaded if this is not specified.}
}
\description{
Opens a .pgen or PLINK 1 .bed file.
}
