% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NewPgenWriter}
\alias{NewPgenWriter}
\title{Creates a new .pgen file for writing.}
\usage{
NewPgenWriter(
  filename,
  sample_ct,
  variant_ct,
  nonref_flags = as.logical(c(TRUE)),
  allele_ct_limit = 2L,
  hardcall_phase_present = FALSE,
  dosage_present = FALSE
)
}
\arguments{
\item{filename}{.pgen file path.}

\item{sample_ct}{Number of samples.}

\item{variant_ct}{Number of variants that will be written.}

\item{nonref_flags}{TRUE when the data is from an ordinary PLINK 1 .bed
file (where the A2 allele is major rather than consistently reference),
FALSE when the REF allele is consistently reference, or a logical vector of
length variant_ct when this is mixed.  NULL is also permitted (this
delegates tracking of nonref information to the .pvar file), but it's
discouraged since .pgen+.bim+.fam is a useful data representation with
direct plink2 support.  Defaults to TRUE, as in the Python PgenWriter.}

\item{allele_ct_limit}{Maximum number of alleles (REF + ALTs) that any
variant can have; defaults to 2.  Values larger than 2 are only useful with
AppendAlleles().}

\item{hardcall_phase_present}{Whether phased hardcalls may be written (with
AppendAlleles()).  Defaults to FALSE.}

\item{dosage_present}{Whether dosages may be written (with
AppendDosages()).  Defaults to FALSE.}
}
\value{
A pgen_writer object.
}
\description{
Only the .pgen is written; you are responsible for writing the companion
.pvar (or .bim) and .psam (or .fam) files yourself, with samples in the
same order as in the vectors passed to the append functions, and variants
in append order.  After all variant_ct variants have been appended (with
AppendBiallelic(), AppendAlleles(), or AppendDosages()), ClosePgenWriter()
must be called to finalize the file.
}
