% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NewPvar}
\alias{NewPvar}
\title{Loads variant IDs and allele codes from a .pvar or .bim file (which can be
compressed with gzip or Zstd).}
\usage{
NewPvar(filename)
}
\arguments{
\item{filename}{.pvar/.bim file path.}
}
\value{
A pvar object, which can be queried for variant IDs and allele
codes.
}
\description{
Loads variant IDs and allele codes from a .pvar or .bim file (which can be
compressed with gzip or Zstd).
}
