% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Read}
\alias{Read}
\title{Loads the variant_numth variant, and then fills buf with numeric dosages
in [0, 2] indicating the dosages of the first ALT (or user-specified)
allele for each sample, with missing values represented by NA.}
\usage{
Read(pgen, buf, variant_num, allele_num = 2L)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}

\item{buf}{Buffer returned by Buf().}

\item{variant_num}{Variant index (1-based).}

\item{allele_num}{Allele index; 1 corresponds to REF, 2 to the first ALT
allele, 3 to the second ALT allele if it exists, etc.  Optional, defaults
to 2.}
}
\value{
No return value, called for buf-filling side-effect.
}
\description{
This function treats the data as diploid; divide by 2 to obtain haploid
dosages.
}
