% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadHardcalls}
\alias{ReadHardcalls}
\title{Loads the variant_numth variant, and then fills buf with \{0, 1, 2, NA\}
values indicating the number of copies of the first ALT (or user-specified)
allele each sample has.}
\usage{
ReadHardcalls(pgen, buf, variant_num, allele_num = 2L)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}

\item{buf}{Buffer returned by Buf() or IntBuf().}

\item{variant_num}{Variant index (1-based).}

\item{allele_num}{Allele index; 1 corresponds to REF, 2 to the first ALT
allele, 3 to the second ALT allele if it exists, etc.  Optional, defaults
to 2.}
}
\value{
No return value, called for buf-filling side-effect.
}
\description{
This function treats the data as diploid; you can divide by 2, and then
treat 0.5 as NA, if it's actually haploid.
}
