% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ReadIntList}
\alias{ReadIntList}
\title{Returns an integer matrix, where rows correspond to variant_subset, columns
correspond to samples, and values are in {0, 1, 2, NA} indicating the
number of hardcall ALT allele copies.  For multiallelic variants, all ALT
alleles are combined.}
\usage{
ReadIntList(pgen, variant_subset)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}

\item{variant_subset}{Integer vector containing 1-based indexes of variants
to load.}
}
\description{
This function treats the data as diploid; you can divide by 2, and then
treat 0.5 as NA, if it's actually haploid.
}
