% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ClosePgen}
\alias{ClosePgen}
\title{Closes a pgen object, releasing resources.}
\usage{
ClosePgen(pgen)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}
}
\value{
No return value, called for side-effect.
}
\description{
Closes a pgen object, releasing resources.
}
