% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ClosePvar}
\alias{ClosePvar}
\title{Closes a pvar object, releasing memory.}
\usage{
ClosePvar(pvar)
}
\arguments{
\item{pvar}{Object returned by NewPvar().}
}
\value{
No return value, called for side-effect.
}
\description{
Closes a pvar object, releasing memory.
}
