% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetAlleleCt}
\alias{GetAlleleCt}
\title{Returns the effective number of alleles for a variant.  Note that if no
pvar was provided to the NewPgen() call, this function may return 2 even at
multiallelic variants, since the .pgen may not store allele-count
information.}
\usage{
GetAlleleCt(pvar_or_pgen, variant_num)
}
\arguments{
\item{pvar_or_pgen}{Object returned by NewPvar() or NewPgen().}

\item{variant_num}{Variant index (1-based).}
}
\value{
max(2, <number of alleles the variant_numth variant is known to
have>).  Note that if no
}
\description{
Returns the effective number of alleles for a variant.  Note that if no
pvar was provided to the NewPgen() call, this function may return 2 even at
multiallelic variants, since the .pgen may not store allele-count
information.
}
