% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetMaxAlleleCt}
\alias{GetMaxAlleleCt}
\title{Returns the maximum GetAlleleCt() value across all variants in the file.}
\usage{
GetMaxAlleleCt(pvar_or_pgen)
}
\arguments{
\item{pvar_or_pgen}{Object returned by NewPvar() or NewPgen().}
}
\value{
Maximum GetAlleleCt() value across all variants.
}
\description{
Returns the maximum GetAlleleCt() value across all variants in the file.
}
