% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetVariantId}
\alias{GetVariantId}
\title{Convert variant index to variant ID string.}
\usage{
GetVariantId(pvar, variant_num)
}
\arguments{
\item{pvar}{Object returned by NewPvar().}

\item{variant_num}{Variant index (1-based).}
}
\value{
The variant_numth variant ID string.
}
\description{
Convert variant index to variant ID string.
}
