% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetVariantsById}
\alias{GetVariantsById}
\title{Convert variant ID string to variant index(es).}
\usage{
GetVariantsById(pvar, id)
}
\arguments{
\item{pvar}{Object returned by NewPvar().}

\item{id}{Variant ID to look up.}
}
\value{
A list of all (1-based) variant indices with the given variant ID.
}
\description{
Convert variant ID string to variant index(es).
}
