% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HardcallPhasePresent}
\alias{HardcallPhasePresent}
\title{Returns whether explicitly phased hardcalls are present.}
\usage{
HardcallPhasePresent(pgen)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}
}
\value{
TRUE if the file contains at least one phased heterozygous
hardcall, FALSE otherwise.
}
\description{
Returns whether explicitly phased hardcalls are present.
}
