% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IntAlleleCodeBuf}
\alias{IntAlleleCodeBuf}
\title{Returns an empty two-row integer matrix that ReadAlleles() can load to.}
\usage{
IntAlleleCodeBuf(pgen)
}
\arguments{
\item{pgen}{Object returned by NewPgen().}
}
\value{
Integer matrix with two rows, and appropriate number of columns for
ReadAlleles().
}
\description{
Returns an empty two-row integer matrix that ReadAlleles() can load to.
}
